refactor merging request labels to tags into a function
