+++ pkg/service/labels/labels_test.go
@@ -1,5 +1,11 @@
 package labels
 
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
 //func TestScopeTag(t *testing.T) {
 //	config.Set(config.Config{Env: "dev"})
 //
@@ -22,3 +28,14 @@ package labels
 //	assert.Equalf(t, expected, got, "DefaultTags: ")
 //
 //}
+
+func TestMergeRequestLabel(t *testing.T) {
+	tags := make(map[string]*string)
+	a := "hello"
+	tags["a"] = &a
+	labels := map[string]string{"a": "hi", "b": "hello"}
+	MergeRequestLabel(tags, labels)
+	assert.Equal(t, "hi", *tags["a"], "expected hi, got %s", *tags["a"])
+	assert.Equal(t, 2, len(tags), "length must be 2, got %d", len(tags))
+
+}
